+++ crates/fuel-core/src/service/adapters/fuel_gas_price_provider/ports.rs
@@ -1,29 +1,10 @@
 use fuel_core_types::fuel_types::BlockHeight;
 
-#[derive(Debug, Clone, Copy)]
-pub struct BlockFullness {
-    used: u64,
-    capacity: u64,
-}
-
-impl BlockFullness {
-    pub fn new(used: u64, capacity: u64) -> Self {
-        Self { used, capacity }
-    }
-
-    pub fn used(&self) -> u64 {
-        self.used
-    }
-
-    pub fn capacity(&self) -> u64 {
-        self.capacity
-    }
-}
-
 pub trait GasPriceAlgorithm {
     fn gas_price(&self, block_bytes: u64) -> u64;
+    fn last_gas_price(&self) -> u64;
 }
 
 pub trait GasPriceEstimate {
-    fn estimate(&self, block_height: BlockHeight) -> u64;
+    fn worst_case_gas_price(&self, block_height: BlockHeight) -> u64;
 }