+++ crates/fuel-core/src/service/adapters/fuel_gas_price_provider/tests.rs
@@ -13,12 +13,13 @@ mod graph_ql_gas_price_estimate_tests;
 
 #[derive(Debug, Clone, Copy)]
 pub struct TestGasPriceAlgorithm {
+    last: u64,
     multiply: u64,
 }
 
 impl Default for TestGasPriceAlgorithm {
     fn default() -> Self {
-        Self { multiply: 2 }
+        Self { last: 100, multiply: 2 }
     }
 }
 
@@ -26,10 +27,14 @@ impl GasPriceAlgorithm for TestGasPriceAlgorithm {
     fn gas_price(&self, block_bytes: u64) -> u64 {
         self.multiply.saturating_mul(block_bytes)
     }
+
+    fn last_gas_price(&self) -> u64 {
+        self.last
+    }
 }
 
 impl GasPriceEstimate for TestGasPriceAlgorithm {
-    fn estimate(&self, _block_height: BlockHeight) -> u64 {
+    fn worst_case_gas_price(&self, _block_height: BlockHeight) -> u64 {
         self.multiply.saturating_mul(10_000_000) // Arbitrary fake bytes
     }
 }