+++ crates/fuel-core/src/service/adapters/fuel_gas_price_provider/tests/graph_ql_gas_price_estimate_tests.rs
@@ -8,7 +8,7 @@ async fn estimate_gas_price__happy_path() {
     let gas_price_provider = build_provider(algo);
 
     // when
-    let expected_price = algo.estimate(next_height);
+    let expected_price = algo.worst_case_gas_price(next_height);
     let actual_price = gas_price_provider.worst_case_gas_price(next_height).await;
 
     // then