+++ crates/fuel-core/src/service/adapters/fuel_gas_price_provider/tests/tx_pool_gas_price_tests.rs
@@ -12,11 +12,10 @@ async fn gas_price__if_requested_block_height_is_latest_return_gas_price() {
     // given
     let algo = TestGasPriceAlgorithm::default();
     let gas_price_provider = build_provider(algo);
-    let bytes = 10;
 
     // when
-    let expected_price = algo.gas_price(bytes);
-    let actual_price = gas_price_provider.gas_price(bytes).await.unwrap();
+    let expected_price = algo.last_gas_price();
+    let actual_price = gas_price_provider.gas_price().await.unwrap();
 
     // then
     assert_eq!(expected_price, actual_price);