+++ crates/services/producer/src/block_producer.rs
@@ -121,21 +121,7 @@ where
 
         let header = self.new_header(height, block_time).await?;
 
-        let consensus_params = self
-            .consensus_parameters_provider
-            .consensus_params_at_version(&header.consensus_parameters_version)?;
-
-        let gas_per_bytes = consensus_params.fee_params().gas_per_byte();
-        let max_gas_per_block = consensus_params.block_gas_limit();
-        let max_block_bytes = max_gas_per_block
-            .checked_div(gas_per_bytes)
-            .ok_or(anyhow!("Failed to calculated max block bytes"))?;
-
-        let gas_price = self
-            .gas_price_provider
-            .gas_price(max_block_bytes)
-            .await
-            .map_err(|e| anyhow!("No gas price found for block {height:?}: {e:?}"))?;
+        let gas_price = self.calculate_gas_price(&header).await?;
 
         let component = Components {
             header_to_produce: header,
@@ -155,6 +141,17 @@ where
 
         debug!("Produced block with result: {:?}", result.result());
         Ok(result)
+    }
+
+    async fn calculate_gas_price(&self, header: &PartialBlockHeader) -> anyhow::Result<u64> {
+        let consensus_params = self
+            .consensus_parameters_provider
+            .consensus_params_at_version(&header.consensus_parameters_version)
+            .unwrap();
+        let gas_per_bytes = consensus_params.fee_params().gas_per_byte();
+        let max_gas_per_block = consensus_params.block_gas_limit();
+        let max_block_bytes = max_gas_per_block / gas_per_bytes;
+        self.gas_price_provider.gas_price(max_block_bytes).await.map_err(|e| anyhow!("No gas price found: {e:?}"))
     }
 }
 
@@ -230,22 +227,12 @@ where
                 .expect("It is impossible to overflow the current block height")
         });
 
-        let header = self.new_header(height, Tai64::now()).await?;
-
-        let consensus_params = self
-            .consensus_parameters_provider
-            .consensus_params_at_version(&header.consensus_parameters_version)?;
-
-        let gas_per_bytes = consensus_params.fee_params().gas_per_byte();
-        let max_gas_per_block = consensus_params.block_gas_limit();
-        let max_block_bytes = max_gas_per_block
-            .checked_div(gas_per_bytes)
-            .ok_or(anyhow!("Failed to calculated max block bytes"))?;
+        let header = self._new_header(height, Tai64::now())?;
 
         let gas_price = if let Some(inner) = gas_price {
             inner
         } else {
-            self.gas_price_provider.gas_price(max_block_bytes).await?
+            self.calculate_gas_price(&header).await?
         };
 
         // The dry run execution should use the state of the blockchain based on the
@@ -253,7 +240,6 @@ where
         // use the same configuration as the last block -> the same DA height.
         // It is deterministic from the result perspective, plus it is more performant
         // because we don't need to wait for the relayer to sync.
-        let header = self._new_header(height, Tai64::now())?;
         let component = Components {
             header_to_produce: header,
             transactions_source: transactions.clone(),